refactor(api): restructure fetching quiz questions logic and improve type definitions
- Delete src/utils.ts and move shuffleArray to src/helpers with an improved algorithm
- Introduce new types: Question, QuestionsState, and Difficulty
- Refactor fetchQuizQuestions to use improved helper and centralized API fetching
- Update hooks and type imports accordingly

BREAKING CHANGE: This update restructures the core fetching logic and redefines several types, which may affect dependent modules and type checks.
