+++ src/api.ts
@@ -1,30 +1,29 @@
-import { shuffleArray } from './utils';
+import { Difficulty } from './types/difficulty.enum';
+import { Question } from './types/question.type';
+import { QuestionsState } from './types/question-state.type';
+import { shuffleArray } from './helpers/shuffleArray.ts';
 
-export type Question = {
-    category: string;
-    correct_answer: string;
-    difficulty: string;
-    incorrect_answers: string[];
-    question: string;
-    type: string;
+// Creating question state by appending shuffled answers
+const createQuestionState = (question: Question): QuestionsState => ({
+    ...question,
+    answers: shuffleArray([...question.incorrect_answers, question.correct_answer]),
+});
+
+// Centralized API fetching implementation
+const fetchQuestionsFromAPI = async (amount: number, difficulty: Difficulty): Promise<Question[]> => {
+    const endpoint = `https://opentdb.com/api.php?amount=${amount}&category=21&difficulty=${difficulty}&type=multiple`;
+    const response = await fetch(endpoint);
+
+    if (!response.ok) {
+        throw new Error('Failed to fetch quiz questions');
+    }
+
+    const data = await response.json();
+    return data.results;
 };
 
-export enum Difficulty {
-    EASY = 'easy',
-    MEDIUM = 'medium',
-    HARD = 'hard',
-}
-
-export type QuestionsState = Question & { answers: string[] };
-
-export const fetchQuizQuestions = async (
-    amount: number,
-    difficulty: Difficulty,
-): Promise<QuestionsState[]> => {
-    const endpoint = `https://opentdb.com/api.php?amount=${amount}&category=21&difficulty=${difficulty}&type=multiple`;
-    const data = await (await fetch(endpoint)).json();
-    return data.results.map((question: Question) => ({
-        ...question,
-        answers: shuffleArray([...question.incorrect_answers, question.correct_answer]),
-    }));
+// Main function to fetch quiz questions and return the processed state
+export const fetchQuizQuestions = async (amount: number, difficulty: Difficulty): Promise<QuestionsState[]> => {
+    const questions = await fetchQuestionsFromAPI(amount, difficulty);
+    return questions.map(createQuestionState);
 };
