+++ src/helpers/shuffleArray.ts
@@ -5,6 +5,10 @@
  * @returns The shuffled array.
  */
 export const shuffleArray = <T>(inputArray: T[]): T[] => {
-    const randomComparison = () => Math.random() - 0.5;
-    return [...inputArray].sort(randomComparison);
+    const array = [...inputArray];
+    for (let i = array.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [array[i], array[j]] = [array[j], array[i]];
+    }
+    return array;
 };
