+++ src/types/difficulty.enum.ts
@@ -1,0 +1,5 @@
+export enum Difficulty {
+    EASY = 'easy',
+    MEDIUM = 'medium',
+    HARD = 'hard',
+}
