+++ src/types/question-state.type.ts
@@ -1,0 +1,3 @@
+import { Question } from './question.type.ts';
+
+export type QuestionsState = Question & { answers: string[] };
