+++ src/types/question.type.ts
@@ -1,0 +1,8 @@
+export type Question = {
+    category: string;
+    correct_answer: string;
+    difficulty: string;
+    incorrect_answers: string[];
+    question: string;
+    type: string;
+};
